Partially implemented soft constraints with PuLP
matcher-softconstraints creates a .mps file containing the data of the model while solving the problem, and matcher-mps reads the model from the .mps file and solves the problem. This allows us to solve the problem multiple times without having the program keep opening the excel files and creating the model all over again.